fastapi/requirements.txt
pybit
fastapi[all]
binance-futures-connector
pyjwt
uvicorn[standard]